+++ ElvUI/Modules/bags/sort.lua
@@ -119,6 +119,12 @@ local inventorySlots = {
 	INVTYPE_TABARD = 25,
 }
 
+local conjured_items = {
+	[5512] = true, -- Healthstone
+	[162518] = true, -- Mystical Flask
+	[113509] = true, -- Conjured Mana Bun
+}
+
 local safe = {
 	[BANK_CONTAINER] = true,
 	[0] = true
@@ -219,6 +225,14 @@ local function DefaultSort(a, b)
 
 	if bagPetIDs[b] then
 		bRarity = 1
+	end
+
+	if conjured_items[aID] then
+		aRarity = -99
+	end
+
+	if conjured_items[bID] then
+		bRarity = -99
 	end
 
 	if aRarity ~= bRarity and aRarity and bRarity then
@@ -563,17 +577,14 @@ function B.Sort(bags, sorter, invertDirection)
 			end
 
 			if not blackListedSlots[bagSlot] then
-				local method, allowPartialMatch
+				local method
 				for _,itemsearchquery in pairs(blackListQueries) do
 					method = Search.Matches
 					if Search.Filters.tipPhrases.keywords[itemsearchquery] then
-						if itemsearchquery == "rel" or itemsearchquery == "reli" or itemsearchquery == "relic" then
-							allowPartialMatch = true
-						end
 						method = Search.TooltipPhrase
 						itemsearchquery = Search.Filters.tipPhrases.keywords[itemsearchquery]
 					end
-					local success, result = pcall(method, Search, link, itemsearchquery, allowPartialMatch)
+					local success, result = pcall(method, Search, link, itemsearchquery)
 					if success and result then
 						blackListedSlots[bagSlot] = result
 						break