+++ ElvUI/Modules/skins/addons/ace3.lua
@@ -3,11 +3,11 @@ local S = E:GetModule('Skins')
 
 --Cache global variables
 --Lua functions
+local _G = _G
 local select = select
 --WoW API / Variables
 local hooksecurefunc = hooksecurefunc
 local CreateFrame = CreateFrame
-local GetCVarBool = GetCVarBool
 
 local RegisterAsWidget, RegisterAsContainer
 local function SkinDropdownPullout(self)
@@ -31,7 +31,6 @@ function S:SkinAce3()
 	if not AceGUI then return end
 
 	local oldRegisterAsWidget = AceGUI.RegisterAsWidget
-	local ColorBlind = GetCVarBool('colorblindmode')
 
 	RegisterAsWidget = function(object, widget)
 		if not E.private.skins.ace3.enable then
@@ -62,7 +61,7 @@ function S:SkinAce3()
 			checkbg:SetTexture('')
 			highlight:SetTexture('')
 
-			if not ColorBlind then
+			if E.private.skins.checkBoxSkin then
 				checkbg.backdrop:SetInside(widget.checkbg, 5, 5)
 
 				check:SetTexture([[Interface\AddOns\ElvUI\media\textures\melli]])
@@ -329,6 +328,11 @@ end
 local function attemptSkin(_, _, addon)
 	if addon == 'ElvUI_Config' then
 		local AceGUI = E.Libs.AceGUI
+		if not AceGUI then
+			AceGUI = _G.LibStub('AceGUI-3.0', true)
+			E.Libs.AceGUI = AceGUI
+		end
+
 		if AceGUI and (AceGUI.RegisterAsContainer ~= RegisterAsContainer or AceGUI.RegisterAsWidget ~= RegisterAsWidget) then
 			S:SkinAce3()
 		end