+++ ElvUI/Modules/unitframes/elements/classbars.lua
@@ -8,6 +8,7 @@ local floor, max = math.floor, math.max
 local find, sub, gsub = string.find, string.sub, string.gsub
 --WoW API / Variables
 local CreateFrame = CreateFrame
+local UnitHasVehicleUI = UnitHasVehicleUI
 local MAX_COMBO_POINTS = MAX_COMBO_POINTS
 
 --Global variables that we don't cache, list them here for mikk's FindGlobals script
@@ -168,7 +169,7 @@ function UF:Configure_ClassBar(frame, cur)
 			end
 		end
 
-		if not frame.USE_MINI_CLASSBAR then
+		if (not frame.USE_MINI_CLASSBAR) and frame.USE_CLASSBAR then
 			bars.backdrop:Show()
 		else
 			bars.backdrop:Hide()
@@ -308,6 +309,7 @@ UF.ToggleResourceBar = ToggleResourceBar --Make available to combobar
 function UF:Construct_ClassBar(frame)
 	local bars = CreateFrame("Frame", nil, frame)
 	bars:CreateBackdrop('Default', nil, nil, self.thinBorders, true)
+	bars:Hide()
 
 	local maxBars = max(UF.classMaxResourceBar[E.myclass] or 0, MAX_COMBO_POINTS)
 	for i = 1, maxBars do
@@ -376,6 +378,12 @@ end
 -------------------------------------------------------------
 -- DEATHKNIGHT
 -------------------------------------------------------------
+local function PostUpdateRunes(self)
+	local useRunes = not UnitHasVehicleUI('player')
+	self:SetShown(useRunes)
+	UF.PostVisibilityRunes(self, useRunes)
+end
+
 function UF:Construct_DeathKnightResourceBar(frame)
 	local runes = CreateFrame("Frame", nil, frame)
 	runes:CreateBackdrop('Default', nil, nil, self.thinBorders, true)
@@ -396,7 +404,7 @@ function UF:Construct_DeathKnightResourceBar(frame)
 		runes[i].bg.multiplier = 0.3
 	end
 
-	runes.PostUpdateVisibility = UF.PostVisibilityRunes
+	runes.PostUpdate = PostUpdateRunes
 	runes.UpdateColor = E.noop --We handle colors on our own in Configure_ClassBar
 	runes:SetScript("OnShow", ToggleResourceBar)
 	runes:SetScript("OnHide", ToggleResourceBar)
@@ -404,7 +412,7 @@ function UF:Construct_DeathKnightResourceBar(frame)
 	return runes
 end
 
-function UF:PostVisibilityRunes(enabled, stateChanged)
+function UF:PostVisibilityRunes(enabled)
 	local frame = self.origParent or self:GetParent()
 
 	if enabled then
@@ -413,14 +421,6 @@ function UF:PostVisibilityRunes(enabled, stateChanged)
 	else
 		frame.ClassBar = "ClassPower"
 		frame.MAX_CLASS_BAR = MAX_COMBO_POINTS
-	end
-
-	if stateChanged then
-		ToggleResourceBar(frame[frame.ClassBar])
-		UF:Configure_ClassBar(frame)
-		UF:Configure_HealthBar(frame)
-		UF:Configure_Power(frame)
-		UF:Configure_InfoPanel(frame, true) --2nd argument is to prevent it from setting template, which removes threat border
 	end
 end
 
@@ -433,7 +433,7 @@ function UF:Construct_AdditionalPowerBar(frame)
 	additionalPower.colorPower = true
 	additionalPower.PostUpdate = UF.PostUpdateAdditionalPower
 	additionalPower.PostUpdateVisibility = UF.PostVisibilityAdditionalPower
-	additionalPower:CreateBackdrop('Default')
+	additionalPower:CreateBackdrop('Default', nil, nil, self.thinBorders, true)
 	UF.statusbars[additionalPower] = true
 	additionalPower:SetStatusBarTexture(E.media.blankTex)