+++ ElvUI/Modules/unitframes/groups/raid40.lua
@@ -56,9 +56,6 @@ function UF:Construct_Raid40Frames()
 	self.Range = UF:Construct_Range(self)
 	self.customTexts = {}
 
-	UF:Update_StatusBars()
-	UF:Update_FontStrings()
-
 	self.unitframeType = "raid40"
 
 	return self